pyproject.toml
[project]
name = "saios-starter"
version = "0.1.0"
description = "3 AI agents that replaced my VA — free starter from SAIOS"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["saios"]

[tool.ruff]
line-length = 100
target-version = "py311"
